requirements.txt
fastapi
uvicorn[standard]

langchain
langchain-core
langchain-openai
langchain-community
langchain-text-splitters

openai
faiss-cpu
pypdf
tiktoken
python-multipart
